requirements.txt
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.10
billiard==3.6.3.0
CacheControl==0.12.6
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.2
django-allauth==0.41.0
django-ckeditor==5.9.0
django-cleanup==5.2.0
django-crispy-forms==1.8.1
django-js-asset==1.2.2
django-mailer==2.0.1
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-widget-tweaks==1.4.5
gunicorn==20.0.4
html5lib==1.0.1
icalendar==4.0.5
idna==2.8
importlib-metadata==1.6.0
ipaddr==2.2.0
jsonfield==3.1.0
kombu==4.6.8
lockfile==0.12.2
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
progress==1.5
psycopg2==2.8.6
pyparsing==2.4.6
python-dateutil==2.8.1
python3-openid==3.1.0
pytoml==0.1.21
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.0
swapper==1.1.2.post1
urllib3==1.25.8
vine==1.3.0
webencodings==0.5.1
zipp==3.1.0
